Adds environment configuration, default to `prod`.
This configuration will be used to check the current environment.
Each environment has its own needs and should not impact behavior of others.

When deploying the application, it shutdowns itself immediately because of `StdIn.readLine()`.
To avoid this behavior in prod but keeping the feature in dev, adds the environment verification.