Adds database evolution system.
The database will need evolutions and the work of the developer is not to manage its update.
Because I am alone, I do not want to manage this during evolutions.

So this system will do the following at system startup:

* Check the last played evolution
* Run evolutions appeared after the last one on the right order
* Update the last played evolution number

This feature could slow the startup but it does not feel like it (local database).