Create the Card and RichContent components
To show the list of theaters and other elements, we choose to display them as cards.
These cards are also used into collections.
Because in cards content should not be too important and shown text be similar to the one written,
create a `RichComponent` which show paragraph and limit the shown text.
This component could be used to render a more complicated text (eg: markdown)

This commit do the following:

* Create the `Card` and `CardCollection` components (ts and scss)
* Create the `RichContent` component (ts and scss)
* Update button rendering and refactor their code