Any user can reserve a seat for a play.
It is the main feature of the application.
Any user can reserve seats once the play is created.

To avoid having strong relation between reservations and theaters,
each time we create or update a play, we duplicate the theater plan.
The fallback is when updating the theater, existing play will not be updated.

This commit is for server-side only.

This commit do the following:

* Create the script to create tables for reservations
* Correct the play service to duplicate the theater on creation and update
* Create the Reservation module
* Plug the module into the application
* Create the `findReservedSeats` action used to check which seats are taken
* Create the `create` action to do the reservation