Any user can reserve a seat for a play.
It is the main feature of the application.
Any user can reserve seats once the play is created.

We directly include the reservation form in the play page to avoid useless navigation.

The created `AsyncPage` is to manage pages initializing their state
in terms of their properties.
Before, when we navigate the application with the same page
but different properties (eg: two plays), the page was not loaded.

This commit is for client-side only.

This commit do the following:

* Create the `AsyncPage` component
* Create the play page
* Create the reservation form
* Plug the new page to the navigation