The admin can create a period of membership
The admin need to be able to manage membership (called *Adhesions*).
Each membership is valid during an amount of period (usually september to june).
Each year, the list of activities can change.
So we create a system of period of membership where the admin can
manage periods and activities.

This commit do the following (server-side only):

* Create the script to create the table `period_adhesion`
* Create the Adhesion module inside User module
* Create the `createPeriod` action
* Introduce the notion of `Period`