The admin can create a period of membership
The admin need to be able to manage membership (called *Adhesions*).
Each membership is valid during an amount of period (usually september to june).
Each year, the list of activities can change.
So we create a system of period of membership where the admin can
manage periods and activities.

This commit do the following (client-side only):

* Create the adhesion module inside user module
* Create the page to create period of adhesions
* Update navigation to create the menu for membership management
* Introduce the generic notion of `Period`
*