Complete definiti model and remove duplicated scala verifications.
To continue the work into transforming the model into Definiti version,
we complete the definiti model.
Most of changes are the use of specific messages.

Also, because aggregate validation is done in definiti part,
we simplify or remove duplicated verification that existed in scala.

We cannot remove all verifications because some of them check the universe.

This commit does the following:

* Add specific messages in definiti model
* Simplify verifications
* `PlayPrice.condition` comes back as a `String` instead of an `Option`
* Remove scala version of adhesion models
* Move `ResultSet` to model into services like other cases
* Add a docker helper to have a contained postgres database